Fix flakes due to writes to test tmp dir while test tmp dir is being removed (#10304)
See discussion here:

https://github.com/buildbuddy-io/buildbuddy/pull/10263

Basically, this removes the directory atomically instead of by
recursively removing it. The conversation is, should we leave the
behavior as-is because it will sometimes cause tests to flake if they
write to the directory after the test is over, which may indicate an
error?

My position is that if we want to test whether something writes after a
test is supposed to be finished, we should explicitly test that
behavior, as opposed to relying on an implementation detail of the
cleanup of the test directory to sometimes cause a race condition that
results in a transient error. If we want to test if the directory has
additional entries added during the cleanup phase (which I am opposed
to; I don't think this is the purpose of the test-filesystem cleanup
phase), then we should get the directory tree, sleep, and then diff the
directory tree we got with the current one, and then atomically remove
the directory.