Allow executor to run under tini again, even with child cgroups enabled (#10384)
Running the executor under tini solves an issue with zombie processes
building up, which is non-ideal for debugging (it causes the `top`
command to show a bunch of extra processes which aren't actually
running, and also makes it consume a lot more CPU since it has to query
procfs for all of the zombie processes, which gets expensive when there
are thousands of zombies). It's also non-ideal in the sense that if the
executor runs for long enough, it will eventually run out of pids
(though we don't see this in our own deployments since we restart the
executor somewhat regularly).

The trick to make this work (which I didn't realize before) is to move
both the executor _and tini_ to the executor child cgroup, _before_ we
delegate controllers. If we try to delegate controllers _then_ move the
executor (with tini also running in the original cgroup), it violates
the "no internal process" constraint, so we get an error.

See
https://docs.kernel.org/admin-guide/cgroup-v2.html#no-internal-process-constraint
(highlighted the key part in bold)

> Non-root cgroups can distribute domain resources to their children
only when they don’t have any processes of their own. In other words,
only domain cgroups which don’t contain any processes can have domain
controllers enabled in their “cgroup.subtree_control” files. [...] Note
that the restriction doesn’t get in the way if there is no enabled
controller in the cgroup’s “cgroup.subtree_control”. This is important
as otherwise it wouldn’t be possible to create children of a populated
cgroup. **To control resource distribution of a cgroup, the cgroup must
create children and transfer all its processes to the children before
enabling controllers in its “cgroup.subtree_control” file.**