invocation_timing_card: name the timing profile json
Add a json extension to the Bazel profile filename.
This will ultimately be added to the Content-Disposition response header
of the download request and let the browser saved the file with the same
name.

When user use `gzip -d` to decompress the `.gz` package, the file will
appear as json automatically instead of a generic blob file.
This automatically enables editor syntax highlight and auto formatter to
work on the file more easily.