invocation_execution: show Execution's primary output
Modify GetExecution RPC to start returning primary output path from
OLAP database.

If the path is available, show the last 2 segments of the path on the UI
while allowing the user to filter based on the output path.

The main motivation of this change is to help user differentiate between
multiple CcCompile executions, which can all share the same mnemonic and
target label, but different primary output.

Similarly, when a test target is sharded, all the sharded executions
will have the same target label - mnemonic(TestRunner) pair, but the
last 2 part of the primary output will be
`../shard_3_of_11/test.outputs` which can be a great way to
differentiate between them.