Optimize the limits for using `Batch{Read|Upload}Blobs` RPCs (#10452)
In dev experiments, reducing these limits so that more blobs use the
bytestream API reduces latency for those blobs by 5-10%. Reducing the
limit further didn't have any additional benefits in my experiments, and
it does create more RPCs. It might be that with more realistic
experiments, an even lower limit would be better, but I'm being somewhat
cautious.

The mechanism here is very similar to
https://github.com/buildbuddy-io/buildbuddy/pull/10392 and
https://github.com/buildbuddy-io/buildbuddy/pull/10354.

For example, in the write path, uploading a 4MB blob using the batch api
follows these sequential steps:
- compress the blob on the client side
- send the whole blob as single payload
- receive the blob on the server, which can take 10-100ms
- decompress the whole blob on the server before writing anything (to
check the digest).
- write the blob.

Using the streaming API allows these steps to be overlapping.
Additionally, it reduces the chances that small blobs are in the same
batch as a single large blob which slows them all down.