CLI: Refactor handling of flag forms in preparation for handling old names (#10397)
This PR generalizes the handling of flag forms by `Option`s so that we
can add support for old names of flags more cleanly and easily. It also
de-duplicates some formatting code, moving that functionality into
`optionBase`, and moves the handling of recognizing which form a flag
takes from the parser and into `optionBase` as well.

It does not yet have tests for `optionBase`, which is why I am
submitting it as a draft. When I've written tests, I will mark it ready
for review.

The principal motivations for this refactor:

- Make it impossible to represent a flag that claims to use, for
example, both a short name and an old name
- Make it easy to add new aliases in the future
- Encapsulate common behavior in the handling of flag formatting
- Encapsulate common behavior in the handling of recognizing formats of
flags passed in the command line or bazelrc
- Avoid potential for adding plugin flags that could collide with
negative forms of flags