[build_event_server_test] pull out test BES server into separate package (#10474)
In order to test the `build_event_publisher.Publisher` interface, it
would help to have a test build event server. We already have one! This
change pulls it out so other tests can use it.

This change also switches `TestBuildEventServer` to use a `bufconn`
instead of a Unix domain socket.

This change is part of a series of changes to add test coverage for
build event publishing and build event reporting in `ci_runner`.