Correctly populate `FAILED_TO_BUILD` status in api (#10493)
Currently we handle test statuses correctly, but assume that if we
receive a `BuildEvent_Completed` that the target built correctly.

This is not accurate, because the Completed event contains a `success`
field that we're currently ignoring.

This PR fixes the API to correctly report the `FAILED_TO_BUILD` status
and adds a test for this.