Allow limiting runner pool size based on total memory (#10518)
Currently we only allow limiting based on a fixed runner count, and we
*implicitly* limit the total max memory usage based on this count
multiplied by the individual runner max memory limit. But in practice,
it's pretty infrequent that we're getting close to this implicit total
memory limit (we have metrics for this), so the current count-based
limit is too coarse-grained.

This PR allows also limiting based on the actual memory usage, as an
alternative to limiting based on count. We still do want to limit based
on count probably, but we can set the limit to be much higher (something
like 100 runners maybe), and we can mostly lean on the total memory
limit instead. This way, we can fit more runners, since some runners use
less memory than others.