replay_invocation: implement raw_json_file
When replay events from a json file using -build_event_json_file flag,
we will attempt to parse the events line-by-line until the end of file
in compliance with the Chrome Trace Event Format(TEF) specs.

However the json file we got from clicking the Download button in the
Invocation page's Raw tab has already been converted from proto to json
and will contain trailing commas in each line.

Introduce a new raw_json_file flag for this specific type of file and
ignore the trailing comma `,` on each line so that we get a complete
json for unmarshalling.