Better fix for tmp file leak (#10538)
While I was falling asleep last night, I realized that the `defer` order
in my first fix was wrong. If we do acquire quota, it's released before
we delete the file. This doesn't affect correctness, but it does mean
that getting the quota has no effect.

Fixed this, added an unexpected event when we fail to delete temp files,
and made `Close` always return nil, since the error returned from
`reserveFileWriterQuota` isn't relevant to the caller.