invocation_exec_log_card: avoid calling GetExecution twice
Currently we are calling GetExecution twice in an interval when the
invocation is still in progress. One is triggered by
fetchUpdatedProgress, which updates the invocation model and trigger
componentDidUpdate to call GetExecution again.

Tighten the componentDidUpdate trigger by making sure that we would only
fetchExecution when the invocation id or status changed.
This leaves the timer-driven polling the only remaining source of
periodic GetExecution requests.