Add metrics for child cgroups (#10546)
Now that we have child cgroups, it would be useful to be able to see the
memory usage of the executor process (+ any other misc child processes
that it spawns, like `ip`, `mkfs.ext4`, etc.) separately from the memory
usage of any task containers. This would give a better idea of how much
memory we need to allocate for the executor itself.

This PR adds a metric for that (on Linux only for now, since it's based
on the child cgroup setup). Later we can add other things like CPU usage
counters if needed.