cli: fix plugin arg file handling (#10567)
When we pass bazel args through the CLI plugin pipeline, we're currently
dropping args that are empty.

This PR fixes that by preserving blank lines in the args file. This does
mean that plugins now have to be more careful about whitespace handling,
which might be annoying, but I'm not sure if there's a better solution
:thinking: