Store pebble cache tmp files in pebble/root/directory/tmp and clean up on server start (#10552)
In
https://buildbuddy-corp.slack.com/archives/C06N9AHA4JX/p1761071985216169
we found that the temporary files created in
[disk.WriteFile](https://github.com/buildbuddy-io/buildbuddy/blob/371c7cf2078e1c5a306892f750fb581628848b77/server/util/disk/disk.go#L84)
were not being removed in some cases due to a context bug. Vanja fixed
this in https://github.com/buildbuddy-io/buildbuddy/pull/10537 and then
fixed it better in
https://github.com/buildbuddy-io/buildbuddy/pull/10538. But it's still
technically possible to leak these files. For example, if the server
shuts down abruptly and the deferred functions don't have a chance to
run. The impact of this problem is smaller, but this PR fixes it by
storing these temporary files in a separate subdirectory (under the
Pebble cache's root directory) so they can easily be found and removed
on server startup.

One possible concern with this approach is that different partitions of
the pebble cache could be on different disks, slowing down the move
operation that's done when the file is committed. In that case, we could
asynchronously scan for and delete these files like we do in
[disk_cache.go](https://github.com/buildbuddy-io/buildbuddy/blob/371c7cf2078e1c5a306892f750fb581628848b77/server/backends/disk_cache/disk_cache.go#L740-L793).
I tested this and estimate it would take about 4 minutes to scan and
clean the on the app's pebble cache on startup.

Fixes: https://github.com/buildbuddy-io/buildbuddy-internal/issues/5868