scim: add support for Entra SCIM multi roles
In addition to `SingleAppRoleAssignment` expression, Azure Entra also
offer 2 other app role expressions which will result in payloads with
multiple roles:
- AppRoleAssignmentsComplex
- AssertiveAppRoleAssignmentsComplex

Add support for these 2 expressions by matching on `roles` attribute
when there are no `primary eq "True"` was matched.

When we got multiple non-primary roles from Entra, we will pick the role
with the higher-privilege to assign to the user. For example, if the
user was given both "Admin" and "Developer" role, the effective role
will be "Admin".