Upgrade clickhouse version in tests (#10420)
Clickhouse 24.8 added a table setting,
`deduplicate_merge_projection_mode`, that is required when adding
projections with merge engines. It's not required on existing
projections, but since tests recreate databases, I needed to added it to
schema.go. This option didn't exist in previous versions, so to make
sure that we don't try to load a snapshot with an old clickhouse
version, I changed the recycling key for all the tests that reuse
clickhouse from snapshots.

I tested that the app is able to start and migrate the DB even with that
setting, because we only add the projection if it doesn't already exist.

Finally, I had to add `CLICKHOUSE_DEFAULT_ACCESS_MANAGEMENT=1` to keep
the legacy behavior of allowing the default user access without a
password.