Support configuring enabled redis shards via experiment config (#10597)
This PR allows dynamically reconfiguring the redis ring client using
experiment configs, so that we can support migrations (e.g. adding
shards) while incurring minimal disruption in service.

The approach in this PR is to provide a custom ConsistentHash function
to the ring which filters out the shards that are not enabled in the
experiment config (or the `default_enabled_shards` flag if the
experiment is not enabled/fails to evaluate).

Example usage:

App config:

```yaml
app:
  default_sharded_redis:
    # All available shards.
    # All of these shards will be added to the ring, but only the ones listed in
    # default_enabled_shards (or the experiment, if configured) will be used.
    shards: ["localhost:6379", "localhost:6380"]
    # Enabled shards to use if the experiment is not yet configured, or fails to evaluate for any reason.
    # If empty, all shards in the `shards` list are assumed to be enabled.
    default_enabled_shards: ["localhost:6379"]
```

Experiment config:

```json
{
  "flags": {
    "app.default_sharded_redis.migration": {
      "state": "ENABLED",
      "variants": {
        "default": {
          "enabled_addrs": ["localhost:8379", "localhost:8380"]
        }
      },
      "defaultVariant": "default"
    }
  }
}
```