[SS] Cache snapshot manifest locally immediately after fetching it (#10104)
Currently this situation can happen: 

1. `pr-1` runs on Executor A. It fetches the remote fallback snapshot on
`main`. When it finishes running, it saves a local snapshot for `pr-1`.
Even though some of the `main` snapshot chunks exist locally, it does
not save a local manifest for `main`
2. The remote `main` snapshot is updated
3. `pr-2` runs on Executor A. It fetches the remote fallback snapshot on
`main`. Unfortunately it's unable to use all the local snapshot chunks
that were fetched during # 1, because the `main` snapshot was changed

With this PR, the following can happen instead: 

1. `pr-1` runs on Executor A. It fetches the remote fallback snapshot on
`main`. It saves that `main` manifest locally
2. The remote `main` snapshot is updated
3. `pr-2` runs on Executor A. Because # 1 saved the `main` manifest
locally, it will use that instead of fetching the remote snapshot. It
will reuse more local snapshot chunks

Should be used with this change:
https://github.com/buildbuddy-io/buildbuddy/pull/10244