action_card: show target label, mnemonic and primary output
We already showing these information from the Executions tab.
It makes sense to also show them in the action detail page.

This is very much a draft status because I don't like the extra
getExecution rpc that we will now have to make despite having the
executeResponse available.

An alternative way would be to pass the label + mnemonic down through
props. And that from the Executions tab and the Cache tab, we can simply
propagate those existing information down.