metacache: optimize setMulti (#10678)
add a fn param to writerWithSizeHint and rename it to
writerWithSizeHintAndCommitFn. In Set and Writer(), we call md server in
Commit immediately. This is the same as current behavior. In SetMulti(),
we will
pass a different fn where we will put the metadata into a slice and
write the
metadatas in one call.

https://github.com/buildbuddy-io/buildbuddy-internal/issues/6005