Add a flag to use async clickhouse inserts
I disabled our Invocation batching when async inserts are on. Our batching shouldn't be necessary, and [could cause back pressure](https://clickhouse.com/blog/asynchronous-data-inserts-in-clickhouse#disadvantage) when used with async inserts.

One of the downsides of async inserts is that we have to enable deduplication, and then [materialized views aren't supported](https://github.com/ClickHouse/ClickHouse/issues/66003), but it seems that this is actively being worked on, and we don't currently use any materialized views.

I also switched tests to use this method.