action_cache_server: record origin of of ActionResult
When user get an AC hit, it might be useful for them to know where did
the cache entry came from.

Today, we already have the worker id recorded within the
ExecutedActionMetadata, but the user don't often care about the worker
alone, they care about the invocation and data attached to that
invocation (i.e. PR number, Commit hash, User, etc...).

So let's start record the origin of the ActionResult based on the
incoming RequestMetadata when it's available. This will be kept within
the ActionResult's auxiliary_metadata for easy keeping.

NOTE: This should not be treated as a security feature since
RequestMetadata can be dupe and modified by custom clients.