update rules_k8s to unbreak `bazel mod deps --lockfile_mode=update` (#10783)
This change updates the rules_k8s commit to point to a fix that leads to
`bazel mod deps --lockfile_mode=update` succeeding instead of failing
with an error.

Not strictly necessary, but working on the dev_qa test targets I
encountered repos that require an update to their lockfile after
injecting the BuildBuddy toolchain dependency and repos that break when
the lockfile is updated!
Figured it's cleanest to have this repo support lockfile updating!