Don't asynchronously refresh remote_crypter encryption keys (#10779)
The asynchronous refresh mechanism doesn't hang on to the user's
credentials, so these requests fail today. We could fix this by hanging
on to them or adding a bypass mechanism for system RPCs like this, but
for now just synchronously refreshing the key should suffice.

This was definitely causing some unauthenticated errors in
`GetEncryptionKey()` RPCs, but I'm not sure it was causing the
"signature is invalid" errors.

Related issues:
https://github.com/buildbuddy-io/buildbuddy-internal/issues/4758