action_card: fetch stderrout and profile on completion
Currently if we open the action page when the action was still in
progress, the client will wait until the action completed before
reloading the Action Result.

Make sure that when we fetch the Action Result, we also fetch the
stderr, stdout and the profile of the remote action. Otherwise user
will have to manually refresh the page to see these information.