Revert "Update images" (#10773)
Reverts buildbuddy-io/buildbuddy#10759

I think this change is probably fine and (I know we've tested it), but
rather than merging directly to master I would like us to build and push
these and mirror some traffic over and make sure it's 100% before we
drop it for everyone. Sorry for not speaking up about this sooner.