[dev qa test] `test //...` in buildbuddy_dev_qa_test (#10811)
This change allows the `buildbuddy_dev_qa_test` target to run tests
across the server, app, and enterprise directory trees. This brings the
coverage of this test on par with the existing `dev_qa.py` test (and on
a much more recent version of the codebase, to boot). Once we have cut a
release that includes this change, I'll point the dev qa workflow at the
//tools/dev_qa_test/... targets.