Add protos for cache service to fetch cache metadata (#10851)
We fetch cache metadata in 2 places:

1. The cache requests card in the UI shows metadata about cache
artifacts (both CAS and AC)
2. The fetch server needs to fetch cache metadata because the remote
asset API only provides the digest hash and not the digest size. We need
the digest size because our bytestream APIs to read and write blobs
expect a fully formed digest including both hash + size. This is
CAS-only

Currently the BuildbuddyService exposes a GetCacheMetadata RPC that
we're using for these cases. When we move to a world where cache routing
is enabled and there's not a single source of truth for the cache, each
cache will need to expose a RPC for cache metadata

In the short-term this is needed to support the remote asset API in the
cache proxies