refactor: split up platform package to break some deps (#10870)
This package splits up the `platform` package into two packages:
- `server/util/platform` now contains the `Properties` struct and
related parsing logic for converting the string properties to the typed
Go struct. It also contains the "enum" constants for various properties.
This package is not under `enterprise/` anymore, so that we can depend
on `platform.Properties` in interfaces.go (this was the main motivation
for this PR). This also breaks a dep from platform -> interfaces, which
previously resulted in the `PoolType` enum having to awkwardly live in
the `interfaces` package instead of in the platform package.
- `enterprise/server/remote_execution/executorplatform` now contains the
executor-specific platform logic and flag definitions, including the
isolation type config. Previously, the app was bringing in a bunch of
executor flag definitions which don't actually do anything. It's also
incorrect to reference any of these flags, since they won't be defined
in the apps.

Th actual changes in this PR are mostly just moving things around and
updating references. The one non-trivial code change is that the
`*enableVFS` flag logic is now moved to
`executorplatform.ApplyOverrides`, since it references an executor flag
(previously, `ParseProperties` would reference an executor flag whenever
we called this function within the apps, which seems wrong).