[oci resolver] always use internal image + layer structs (#10869)
This change causes the OCI Resolver to always use our internal
`imageFromRawManifest` and `layerFromDigest` types. In practice,
BuildBuddy prod executors have followed this path for months. On-prem
customers have likely not taken this path. Simplifying the code reduces
the number of call sites that will have to switch to the
OCIFetcherClient.

This change also removes some unused flags that were part of the OCI
caching rollout.