Patch gRPC to have more tiers in the default buffer pool (#10873)
This improves bytestream reads and writes by 10-20% in benchmarks, while
allocating 10% less.

With the previous tiers, payloads of 32KiB+1byte would fall in the 1MiB
tier, so they would allocate a lot more memory, plus every time we would
get a 1MiB from the pool, we would clear the whole thing, even though we
only use a part of it.

I'll send a PR or issue in the grpc-go repo.