refactor: improve task sizer interface (#10874)
Instead of having `TaskSizer` take an `ExecutionTask` in its interface
methods, have it take the action, command, and parsed platform props
explicitly (as needed).

This has a few benefits:
- It allows `Update` to call `Get`, which should allow us to experiment
with task sizing logic that updates the current task size based on the
previous size in various ways - e.g. not changing the size too
drastically relative to the previous size
- It lets us avoid re-parsing platform props, and simplifies the error
handling associated with that.
- It reduces dependencies, since ExecutionTask is a big proto with lots
of other fields that aren't needed for task sizing.