filecache: fix failing to rm temp file
Our user got this error when running Windows executor

  Failed to remove filecache temp file: remove C:\<some-path>.tmp: The process cannot access the file because it is being used by another process.

It seems like we need to make sure to close the file handle before
deleting the temp file. Switch to os.WriteFile instead to get it closed
automatically.

Add a test to validate the fix. Worth noting that we are not currently
running filecache test regularly on Windows CI. That should be fixed in
a future PR.