disk_windows: fix windows file size estimation
On Linux, we use this

  ```
  os.FileInfo.Sys().(*syscall.Stat_t).Blocks * 512
  ```

To estimate the impact storing the file would have on the current disk.
It counts the block allocated by the file system and multiply it by the
block size (assume to be 512 bytes).

Add similar estimation for Windows based on cluster allocation.
We picked 4KiB based on the default NTFS and ReFS(DevDrive) block size.

References:
- https://techcommunity.microsoft.com/blog/filecab/cluster-size-recommendations-for-refs-and-ntfs/425960
- https://learn.microsoft.com/en-us/windows-server/storage/file-server/ntfs-overview#support-for-large-volumes