[FC] Take 2- Don't save snapshots when the balloon is malfunctioning (#10947)
Initial attempt
[here](https://github.com/buildbuddy-io/buildbuddy/pull/10850/changes)

In GCP, the balloon can be slow to inflate (maybe due to nested
virtualization? This pattern doesn't appear in SJC. Curious to see if
this [PR](https://github.com/buildbuddy-io/buildbuddy/pull/10946) helps
with that), but that doesn't imply corruption. Thus the previous
iteration caused us to throw lots of good snapshots away

In this PR, only throw snapshots away if deflation fails. That's both a
signal of snapshot corruption, and may prevent future workloads from
failing, if not enough memory is available to them