[FC] Decrease memory threshold for corrupted balloon state (#10979)
I realized the memory threshold of 70% is too high for whether to
consider a balloon faulty.

Let's say availableMem = 1000. The balloon target is 80% of that, so
balloonTarget = 800. Then lets say the balloon stalls at 80% of the
balloon Target: 800 * .8 = 640

At this point the runner is in an unhealthy state, with 64% of its
memory unusable by future workloads because it's being occupied by the
balloon. But we won't trigger the if statement that returns the error,
because balloon size 640 is only 64% of the available memory, and
doesn't hit the 70% threshold