Proactive cancellation (#10940)
Add app and executor flags to enable proactive cancellation - removing
tasks from executor queues as soon as they've been completed by another
executor. This has two benefits:
- It should help reduce unnecessary scale-ups by evicting things from
the queue sooner, making the queue length metric more accurate.
- It should help reduce long-tail queue times, since there are cases
where large tasks are at the head of the queue, blocking other smaller
tasks from running, even though the large blocking task has already been
completed.
- It should help a bit with scheduling latency, since the scheduling
loop doesn't have to waste as many round trips to the app, churning
through tasks that have already been completed. For example, if there
are N completed tasks in the queue, followed by 1 task that hasn't been
completed yet, we have to do N round trips to the app (churning through
the completed tasks) before we can run that 1 task. In the worst case,
all of those round trips happen in serial.
 
Note that we do have "queue pruning" already, which already gives us
some of these benefits - but the benefits are limited, for two reasons:
- It's polling based - in the worst case, we can spend up to 15s waiting
on a task that has already completed by another executor.
- It only looks at the head of the queue (to avoid the excessive
overhead of checking for each task's completion). Pruning tasks only
from the head of the queue is fine if we only care about the throughput
benefits - but we want to prune tasks everywhere in the queue in order
to maximize the autoscaling benefits.

The overall approach is to store redis entries mapping task IDs to the
set of executor IDs which have ever received a probe for the task. When
the task lease is finalized and we delete the task, we also send
proactive cancellation requests to each executor which may have received
a probe.

With this PR, we only trigger proactive cancellation when tasks
complete, but ideally we'd trigger the cancellation even sooner, when
the lease is first claimed. This should be fine in theory, but it seems
slightly riskier, so we can save that for later.

The term "proactive cancellation" was used in the original Sparrow paper
which our scheduler is based on:
https://people.eecs.berkeley.edu/~matei/papers/2013/sosp_sparrow.pdf