[FC] Use /proc/meminfo to determine available memory for the balloon (#10958)
Previously we explicitly dropped the page cache in the guest and used
the balloon statistics to determine how much memory to reclaim. This
approach could be increasing memory thrash, increasing the likelihood of
a stall and guest corruption.

When the cache is dropped, the kernel must synchronously reclaim the
page cache, which incurs some CPU and memory overhead. Then immediately
afterwards, we are inflating the balloon, which allocates a lot of
memory. This is also a CPU intensive process and requires a lot of
memory accounting. These two processes happening simultaneously could
compete and cause pressure on the guest

The balloon is able to naturally reclaim the page cache, so it doesn't
have to be explicitly freed. This way, the memory accounting overhead
only happens once, and because we're only reclaiming ~80% of the cache,
the kernel can preserve what it perceives to be the most valuable cached
data