Implement CAS Split and Splice APIs (#10746)
An implementation for CAS/SplitBlob and SpliceBlob is described here:
https://github.com/bazelbuild/remote-apis/pull/353

In order to do chunking on a layer above the abstraction of the cache,
and store individual chunks in the CAS separately, we need to create a
chunked metadata storage of some sort. To keep things simple, this
implementation uses the Action Cache for storage of the chunked
manifests, and stores them **under the original blobs digest**. This is
simpler than using a **derived digest** (i.e. a hash of the digest +
metadata) but has the same security. The AC entries are stored under a
versioned prefix in the instance name, which means we can change the
version to invalidate all cached manifests across all instances.

`Split`: is used to retrieve a chunked manifest for a blob. If any of
the chunks are not found, or the manifest is not found, it returns a
`NotFound` error.

`Splice`: is used to upsert a chunked manifest. All chunks should be
available in the CAS. If any are missing, it returns an
`InvalidArgument` error. Splice will also return an `InvalidArgument`
error if the chunks do not concatenate together to equal the original
blob digest.

The implementation uses the experiment config so that we can enable this
by group or user gradually to evaluate performance. Since AC entries are
stored by group, this is safe.

This PR is part of a series for an MVP of CDC. The next step is to
implement the `Read`/`Write` ByteStream APIs to read and write using CDC
if a blob matches conditions.

An example of how this can be used for `ByteStream/Read`:
https://github.com/buildbuddy-io/buildbuddy/pull/10997

Follow-ups included in
https://github.com/buildbuddy-io/buildbuddy-internal/issues/6426